+++ client/src/App.js
@@ -2,12 +2,15 @@ import './App.css';
 import ContactUs from './pages/ContactUs';
 
 import DeveloperLogin from './pages/DeveloperLogin';
+import CustomerHome from './pages/CustomerHome';
 
 function App() {
   return (
     <>
+    <CustomerHome />
     <DeveloperLogin />
     <ContactUs />
+
     </>
 
   );
