+++ client/src/components/CustomerStoreList/CategoryFilter/index.js
@@ -1,9 +1,9 @@
 import React, { useEffect } from "react";
-import { useQuery } from '@apollo/react-hooks';
-import { QUERY_CATEGORIES } from "../../utils/queries";
-import { UPDATE_CATEGORIES, UPDATE_CURRENT_CATEGORY } from "../../utils/actions";
-import { idbPromise } from "../../utils/helpers";
-import { useStoreContext } from '../../utils/GlobalState';
+import { useQuery } from '@apollo/client';
+import { QUERY_CATEGORIES } from "../../../utils/queries";
+import { UPDATE_CATEGORIES, UPDATE_CURRENT_CATEGORY } from "../../../utils/actions";
+import { idbPromise } from "../../../utils/helpers";
+import { useStoreContext } from '../../../utils/GlobalState';
 
 function CategoryFilter() {
   const [state, dispatch] = useStoreContext();
