+++ client/src/utils/GlobalState.js
@@ -6,11 +6,7 @@ const { Provider } = StoreContext;
 
 const StoreProvider = ({ value = [], ...props }) => {
   const [state, dispatch] = useProductReducer({
-    products: [],
-    cart: [],
-    cartOpen: false,
-    categories: [],
-    currentCategory: '',
+    stores: [],
   });
 
   return <Provider value={[state, dispatch]} {...props} />;
